Update README(s) a bit
Remove some things that have been done,
add some things that need to be done,
describe some things a bit.

Also 'fix' reading local markdown files in Firefox.
This fix was already there but didn't work except with
another Firefox extension, and it wasn't checked in.
So I added it and describe that.