Poor man's age decryption 'agent' integrated with our desktop environment
It's clunky but it works:

1. unlock age session with meta-W
2. now meta-w can decrypt age-encrypted passwords

Status is not shown correctly yet in status bar.

There's a lot of shuffling with environment variables
to prevent a long-running session id being visible in
any process's environment variables.

This should be enough to prevent:

- harvesting of the age key by naive programs
  (i.e. programs without knowledge about how
  our 'agent' works; yes, this is a bit of
  security through obscurity)
- our age key being stored on disk unencrypted

It's NOT enough to protect against malicious
programs in the user's session, but that
would be impossible.