Only git annex enableremote if remote is not known yet
Earlier I made a mistake with the credentials, causing a failed enableremote attempt every minute.
Each attempt made a commit on the git-annex branch with an updated timestamp.
I'd like to prevent an enormous amount of git commits when such a problem occurs in the future
and I don't spot it right away.

This might do the trick. Or not. To be seen.